Refactor 05-x11-text-cursor-hello
Change the long if-else if statement into into a switch-case statement
in order to make it generally more efficient.
Also remove a superfluous namespace.
Also fix an issue when pasting text while text is selected. Now, delete
the selected text prior to inserting text, giving the impression the
pasted overwrites the seleced text.