Rewrite Cmake setup for yaml-cpp-hello
Apparently, a normal setup with ExternalProject_Add() works only with
make but fails with Ninja. This is a known issue that may be resolved in
later CMake versions. However, I run CMake 3.10 which appears to be the
default version on Ubuntu 18. Use ExternalProject_Add() merely to
download the sources and configure/build the project in a separate file
yaml-cpp.cmake.

This works both with make and Ninja. I like this approach
since I have complete control of the configuration & build of the
external dependency and could nicely source out the yaml-cpp related
configurations in order to increase readability of the main
CMakeLists.txt and allow for easier reausability.

There might be alternative CMake commands to ExternalProject_Add() that
might be better suited for downloading git repositories. I will consider
them after I found any.