Fix broken x11 text widget
Apparently, there were issues when I delete a line or lines both in a
dynamized and in an old version. The vector class does some
deletions and reallocations when adding and removing lines. Apparently,
since I use emlace, they use move constructors and move assinment
operators as well as copy constructors for that. However, moving the raw
pointers buf in the lines and then deleting the old objects invalidates
the buffers. Fix it by copying the buffers also in the move
constructors and assignment operators.

Also add a copy constructor in case someone will use vector::push_XXX()