Add adjust y offset logic in x11 text widget
Add functionality to have the cursor always visible when moving the
cursor vertically.

Change the text widget size to showcase the new behavior.

Move the cursor pos calculation into a separate function since its
coordinates are also necessary for the offset calculation.

Consequently, in order to avoid narrowing conversions, change the
according types of the members x and y.